Add zoom simulation to Dataset Generator
The new -z (--zoom_level) flag is used for creating drop points
of sizes that are accurate to what the camera will be picking up
at different zoom levels. It takes into account the resolution
of the generated images and adjusts the --scale and
--scale_target flags based off of a linear relationship that
I found to be working for the time being.